+++ src/mass_driver/process_repo.py
@@ -30,10 +30,11 @@ def clone_repo(
 ) -> tuple[ClonedRepo, GitRepo]:
     """Clone a repo (if needed) and switch branch"""
     repo_gitobj = clone_if_remote(repo.clone_url, cache_path, logger=logger)
-    switch_branch_then_pull(repo_gitobj, repo.force_pull, repo.upstream_branch)
+    commit = switch_branch_then_pull(repo_gitobj, repo.force_pull, repo.upstream_branch)
     cloned_repo = ClonedRepo(
         cloned_path=repo_gitobj.working_dir,
         current_branch=repo_gitobj.active_branch.name,
+        commit_hash=commit,
         **repo.model_dump(),
     )
     return cloned_repo, repo_gitobj