+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The project uses semantic versioning (see [semver](https://semver.org)).
 - New `ExceptionRecord` class, used in field `PatchResult.error`, captures
   exceptions found during execution, while remaining serializable.
 - New `Error` class for capturing for each repo, what went wrong in any phase
+- `ClonedRepo` now includes `commit_hash` string, to know what you cloned
 - Add `GlobFileEditor.before_run()` to run arbitrary code before processing the
   first file
 - New function `replace_many` in `patchdrivers.brick`, to replace many patterns